refactor:changed type of id to unsigned big integer in foreign key fields

+++ database/migrations/2023_06_12_095750_create_comments_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
             $table->text('content');
-            $table->unsignedInteger('post_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('post_id')->references('id')->on('posts');
             $table->foreign('user_id')->references('id')->on('users');
             $table->timestamps();
